+++ README.md
@@ -10,6 +10,25 @@ A Prometheus exporter for macOS system metrics using `powermetrics` and `vm_stat
 - **Page Size**: Dynamic page size detection
 - **Prometheus Format**: Native Prometheus metrics format
 - **LaunchDaemon Support**: Automatic startup with macOS as root service
+- **Modular Architecture**: Clean separation of concerns with internal packages
+
+## Architecture
+
+The application is structured with a modular architecture:
+
+```
+├── cmd/
+│   └── main.go                    # Application entry point
+├── internal/
+│   ├── collector/
+│   │   ├── powermetrics.go        # PowerMetrics collector
+│   │   └── vmstat.go              # VM statistics collector
+│   ├── config/
+│   │   └── config.go              # Configuration management
+│   └── server/
+│       └── server.go              # HTTP server and metrics endpoint
+└── test/e2e_test.go               # End-to-end tests
+```
 
 ## Prerequisites
 
@@ -30,7 +49,7 @@ cd mac-powermetrics-exporter
 
 2. Build the binary:
 ```bash
-go build -o mac-powermetrics-exporter
+go build -o mac-powermetrics-exporter cmd/main.go
 ```
 
 3. Install the binary:
@@ -48,6 +67,11 @@ Run the exporter manually (requires sudo):
 sudo ./mac-powermetrics-exporter
 ```
 
+Or run directly from source:
+```bash
+sudo go run cmd/main.go
+```
+
 The exporter will start an HTTP server on port 9127. Access metrics at:
 ```
 http://localhost:9127/metrics
@@ -57,33 +81,39 @@ http://localhost:9127/metrics
 
 The exporter runs as a LaunchDaemon with root privileges to access `powermetrics` without additional sudo configuration.
 
-1. Copy the plist file to LaunchDaemons directory:
+1. Update the plist file to point to the correct binary location:
+```bash
+# Edit ninja.oppai.mac-powermetrics-exporter.plist if needed
+# Ensure the ProgramArguments points to /usr/local/bin/mac-powermetrics-exporter
+```
+
+2. Copy the plist file to LaunchDaemons directory:
 ```bash
 sudo cp ninja.oppai.mac-powermetrics-exporter.plist /Library/LaunchDaemons/
 ```
 
-2. Set proper permissions:
+3. Set proper permissions:
 ```bash
 sudo chown root:wheel /Library/LaunchDaemons/ninja.oppai.mac-powermetrics-exporter.plist
 sudo chmod 644 /Library/LaunchDaemons/ninja.oppai.mac-powermetrics-exporter.plist
 ```
 
-3. Load the LaunchDaemon:
+4. Load the LaunchDaemon:
 ```bash
 sudo launchctl load /Library/LaunchDaemons/ninja.oppai.mac-powermetrics-exporter.plist
 ```
 
-4. Verify the service is running:
+5. Verify the service is running:
 ```bash
 sudo launchctl list | grep mac-powermetrics-exporter
 ```
 
-5. To unload the service:
+6. To unload the service:
 ```bash
 sudo launchctl unload /Library/LaunchDaemons/ninja.oppai.mac-powermetrics-exporter.plist
 ```
 
-6. To start/stop the service manually:
+7. To start/stop the service manually:
 ```bash
 # Start
 sudo launchctl start ninja.oppai.mac-powermetrics-exporter
@@ -91,6 +121,28 @@ sudo launchctl start ninja.oppai.mac-powermetrics-exporter
 # Stop
 sudo launchctl stop ninja.oppai.mac-powermetrics-exporter
 ```
+
+## Development
+
+### Running Tests
+
+Run the end-to-end tests to verify functionality:
+```bash
+go test -v
+```
+
+Run tests for all packages:
+```bash
+go test -v ./...
+```
+
+### Project Structure
+
+- **`cmd/main.go`**: Application entry point that initializes configuration and starts the server
+- **`internal/config/`**: Configuration management with default values
+- **`internal/collector/`**: Metric collectors for powermetrics and vm_stat
+- **`internal/server/`**: HTTP server setup and Prometheus metrics endpoint
+- **`e2e_test.go`**: End-to-end tests that verify the complete application functionality
 
 ## Metrics
 
@@ -179,15 +231,33 @@ max(powermetrics_cpu_active_residency_percent)
 
 ### Port Configuration
 
-To change the default port (9127), modify the `main.go` file:
+To change the default port (9127), modify the `internal/config/config.go` file:
 
 ```go
-log.Fatal(http.ListenAndServe(":YOUR_PORT", nil))
+func New() *Config {
+	return &Config{
+		Port: ":YOUR_PORT",
+	}
+}
 ```
 
 ### Sampling Interval
 
-The exporter uses a 1-second sampling interval for `powermetrics`. To modify this, change the `-i` parameter in the `powermetrics` command within the code.
+The exporter uses a 1-second sampling interval for `powermetrics`. To modify this, change the `-i` parameter in the `powermetrics` command within the `internal/collector/powermetrics.go` file.
+
+### Adding New Collectors
+
+To add new metric collectors:
+
+1. Create a new collector in `internal/collector/`
+2. Implement the `prometheus.Collector` interface
+3. Register the collector in `internal/server/server.go`
+
+Example:
+```go
+// In internal/server/server.go
+prometheus.MustRegister(collector.NewYourNewCollector())
+```
 
 ## Troubleshooting
 
@@ -196,6 +266,7 @@ The exporter uses a 1-second sampling interval for `powermetrics`. To modify thi
 1. **Permission Denied**: Ensure sudo permissions are configured correctly for `powermetrics`
 2. **Command Not Found**: Verify `powermetrics` is available (should be on all modern macOS systems)
 3. **High CPU Usage**: Consider increasing the sampling interval if the exporter consumes too many resources
+4. **Build Errors**: Ensure Go modules are properly initialized with `go mod tidy`
 
 ### Logs
 
@@ -212,6 +283,22 @@ Test the exporter manually:
 curl http://localhost:9127/metrics
 ```
 
+Run the end-to-end tests:
+```bash
+go test -v -run TestE2E
+```
+
+### Debugging
+
+For development and debugging:
+```bash
+# Run with verbose logging
+sudo go run cmd/main.go
+
+# Check if collectors are working
+curl -s http://localhost:9127/metrics | grep -E "(powermetrics|vmstat)" | head -10
+```
+
 ## Security Considerations
 
 - The exporter runs as root via LaunchDaemon to access `powermetrics`
@@ -219,20 +306,31 @@ curl http://localhost:9127/metrics
 - Restrict network access to the metrics endpoint (consider firewall rules)
 - Monitor system logs for service activity
 - The service automatically restarts if it crashes (KeepAlive=true)
+- Internal packages are not exposed externally, following Go best practices
 
 ## Contributing
 
 1. Fork the repository
 2. Create a feature branch
 3. Make your changes
-4. Add tests if applicable
-5. Submit a pull request
+4. Add tests if applicable (especially for new collectors)
+5. Run tests: `go test -v ./...`
+6. Submit a pull request
+
+### Code Structure Guidelines
+
+- Keep collectors in `internal/collector/`
+- Configuration changes go in `internal/config/`
+- Server modifications in `internal/server/`
+- Add end-to-end tests for new functionality
+- Follow Go naming conventions and add appropriate documentation
 
 ## License
 
-[Add your license information here]
+MIT
 
 ## Acknowledgments
 
 - Built with [Prometheus Go client library](https://github.com/prometheus/client_golang)
-- Uses macOS `powermetrics` and `vm_stat` system utilities 
+- Uses macOS `powermetrics` and `vm_stat` system utilities
+- Follows Go project layout standards
